Major rework of sharding on the routing layer
We now have Vshards (which are still expensive to reshard), but these
are then mapped to datastore_shards. These datastore_shards are then
mapped to datasource_instances which are in turn mapped to
storage_nodes. This means we can easily add shards up to the number of
vshards before having to do full collection scans