Add "PrimaryKey" to collection
This allows us to define primary keys per-collection, meaning we don't
have to require "_id" anymore, as each collection defines its pkey.

In conjunction with this it changes the format of queries for get() /
delete() to now take a map of fields, which must be a superset of the
shard-key and the primarykey.