Major rework of join to new format
The old format had a few issues:
    - only allowed for forward joins (specifically it only allowed for
sub-objects)
    - replaced the value in the return structure, requiring that the
caller know the context of the query when they access data
    - only allowed for a single object in the join relation

The new format addresses these issues. A join field is not a list of
strings with the format "<COLLECTION>.field.or.subfield.name" (with an
allowed shorthand of ".field.in.the.collection" (where the collection is
assumed to be the collection the query is against). The resulting
structure then changes to embed the joined objects as a list under the
relation field with a "." appended.

Note: this is a backwards-incompatible change which *will* require join
queries to be reworked.