Change HTTP api to be more RPC
The old wire protocol was basically a single POST with a blob that
looked like:

```
{
    "insert": {
        "db": ...
    }
}
```

This was a holdover from the old API (that supported parallelism etc.).
Since the API is now simpler there is no need to have all the wrapping.
So instead we now have endpoints per-querytype (/raw/insert etc.).