Restructure of sourcecode layout
This shuffles all of the code based on the new desired layout. Previously I had left the source in a `src` directory with the intention of having lots of clients etc. all in this repo-- but after working with it more that seems unweildy. So, to simplify things we'll make this more like a normal golang project (if there is a normal, no one seems to follow the "standard" ;) ). So for this we'll put all CLIs in cmd, tests in test, docs in doc, and the rest of the go packages in the root of the repo